Fucking unicode
turns out discord strips this newfangled "variation" data from unicode characters in channel names so ❄️(emoji version) becomes ❄(the pictogram) which is inconsistently either shown as an emoji or a pictogram. On windows and mac it shows as pictogram but on ios and android it shows as color emoji. :FUUUU: